finished information endpoint

+++ models/file.go
@@ -2,6 +2,6 @@ package models
 
 type File struct {
 	FilePath string  `json:"file_path"`
-	Size     float32 `json:"size"`
+	Size     float64 `json:"size"`
 	MimeType string  `json:"mimetype"`
 }
